requirements.txt
python-telegram-bot[job-queue]==20.7
SQLAlchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1
httpx==0.26.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
alembic==1.12.1